HackLuCTF/soulsplitter/files/requirements.txt
pyzbar
pillow
opencv-python
numpy
